+++ geotrek/trekking/parsers.py
@@ -14,7 +14,6 @@ from geotrek.common.models import Label, Theme
 from geotrek.common.parsers import (
     ShapeParser, AttachmentParserMixin, GeotrekParser, RowImportError, Parser, ApidaeBaseParser
 )
-from geotrek.common.utils.translation import get_translated_fields
 from geotrek.trekking.models import OrderedTrekChild, POI, Service, Trek, DifficultyLevel, TrekNetwork, Accessibility
 
 
@@ -247,7 +246,71 @@ class ApidaeTranslatedField:
         return rv
 
 
-class ApidaeTrekParser(AttachmentParserMixin, ApidaeBaseParser):
+class ApidaeBaseTrekkingParser(ApidaeBaseParser):
+    """Add the `expand_translations` field option.
+    Map an APIDAE translated field as src to a Geotrek translated field as dst in the parser's `fields` attribute. Set
+    the `expand_translations` option to True on that field for the corresponding mapping to be expanded into a mapping
+    of all translation sub-fields for all configured languages.
+
+    For instance:
+
+    fields = {
+        'name': 'nom',
+    }
+
+    turns into
+
+    fields = {
+        'name_fr': 'nom.libelleFr',
+        'name_en': 'nom.libelleEn',
+        'name_es': 'nom.libelleEs',
+        'name_it': 'nom.libelleIt',
+    }
+    """
+
+    apidae_translation_prefix = 'libelle'
+
+    def __init__(self, *args, **kwargs):
+        self._expand_fields_mapping_with_translation_fields()
+        super().__init__(*args, **kwargs)
+
+    def _expand_fields_mapping_with_translation_fields(self):
+        self.fields = self.fields.copy()
+        translated_fields_to_expand = [
+            field for field, options in self.field_options.items()
+            if options.get('expand_translations') is True
+        ]
+        for translated_field in translated_fields_to_expand:
+            src = self.fields[translated_field]
+            del self.fields[translated_field]
+            for lang in settings.MODELTRANSLATION_LANGUAGES:
+                self.fields[f'{translated_field}_{lang}'] = f'{src}.{self.apidae_translation_prefix}{lang.capitalize()}'
+
+    @classmethod
+    def _get_default_translation_src(cls):
+        return cls.apidae_translation_prefix + settings.MODELTRANSLATION_DEFAULT_LANGUAGE.capitalize()
+
+
+def _prepare_attachment_from_apidae_illustration(illustration, translation_src):
+
+    def get_translation_value_of(key):
+        translated_field = illustration.get(key)
+        if not translated_field:
+            return ''
+        return translated_field.get(translation_src, '')
+
+    legende = get_translation_value_of('legende')
+    copyright = get_translation_value_of('copyright')
+    title = get_translation_value_of('nom')
+    return (
+        illustration['traductionFichiers'][0]['url'],
+        legende or title,
+        copyright,
+        title,
+    )
+
+
+class ApidaeTrekParser(AttachmentParserMixin, ApidaeBaseTrekkingParser):
     model = Trek
     eid = 'eid'
     separator = None
@@ -478,25 +541,10 @@ class ApidaeTrekParser(AttachmentParserMixin, ApidaeBaseParser):
         'libelleFr': 'Cet itinéraire n\'est pas balisé',
         'libelleEn': 'This trek is not marked',
     }
-    apidae_translation_prefix = 'libelle'
 
     def __init__(self, *args, **kwargs):
-        self._translated_fields = [field for field in get_translated_fields(self.model)]
-        self._expand_fields_mapping_with_translation_fields()
         self._related_treks_mapping = defaultdict(list)
         super().__init__(*args, **kwargs)
-
-    def _expand_fields_mapping_with_translation_fields(self):
-        self.fields = self.fields.copy()
-        translated_fields_to_expand = [
-            field for field, options in self.field_options.items()
-            if options.get('expand_translations') is True
-        ]
-        for translated_field in translated_fields_to_expand:
-            src = self.fields[translated_field]
-            del self.fields[translated_field]
-            for lang in settings.MODELTRANSLATION_LANGUAGES:
-                self.fields[f'{translated_field}_{lang}'] = f'{src}.{self.apidae_translation_prefix}{lang.capitalize()}'
 
     def apply_filter(self, dst, src, val):
         val = super().apply_filter(dst, src, val)
@@ -623,17 +671,13 @@ class ApidaeTrekParser(AttachmentParserMixin, ApidaeBaseParser):
         illustrations = val
         rv = []
         for illustration in illustrations:
-            files_metadata_list = illustration['traductionFichiers']
-            if not ApidaeTrekParser._is_still_publishable_tomorrow(illustration) or not files_metadata_list:
+            if (
+                    not ApidaeTrekParser._is_still_publishable_tomorrow(illustration)
+                    or not illustration.get('traductionFichiers')
+            ):
                 continue
-            first_file_metadata = files_metadata_list[0]
             rv.append(
-                (
-                    first_file_metadata['url'],
-                    illustration['legende'][translation_src],
-                    illustration['copyright'][translation_src],
-                    illustration['nom'][translation_src],
-                )
+                _prepare_attachment_from_apidae_illustration(illustration, translation_src)
             )
         return rv
 
@@ -920,10 +964,6 @@ class ApidaeTrekParser(AttachmentParserMixin, ApidaeBaseParser):
         else:
             return None
 
-    @classmethod
-    def _get_default_translation_src(cls):
-        return cls.apidae_translation_prefix + settings.MODELTRANSLATION_DEFAULT_LANGUAGE.capitalize()
-
 
 class ApidaeReferenceElementParser(Parser):
 
@@ -997,3 +1037,71 @@ class ApidaeTrekAccessibilityParser(ApidaeReferenceElementParser):
     model = Accessibility
     element_reference_ids = ApidaeTrekParser.natures_de_terrain_ids_as_accessibilities
     name_field = 'name'
+
+
+class ApidaePOIParser(AttachmentParserMixin, ApidaeBaseTrekkingParser):
+    model = POI
+    eid = 'eid'
+    separator = None
+
+    # Parameters to build the request
+    api_key = None
+    project_id = None
+    selection_id = None
+    size = 20
+    skip = 0
+    responseFields = [
+        'id',
+        'nom',
+        'presentation',
+        'localisation',
+        'informationsPatrimoineCulturel',
+        'illustrations',
+    ]
+    locales = ['fr', 'en']
+
+    # Fields mapping
+    fill_empty_translated_fields = True
+    fields = {
+        'name': 'nom',
+        'description': 'presentation.descriptifCourt',
+        'geom': 'localisation.geolocalisation.geoJson',
+        'eid': 'id',
+        'type': 'type',
+    }
+    natural_keys = {
+        'type': 'label',
+    }
+    field_options = {
+        'type': {'create': True},
+        'name': {'expand_translations': True},
+        'description': {'expand_translations': True},
+    }
+    non_fields = {
+        'attachments': 'illustrations'
+    }
+
+    def filter_type(self, src, val):
+        type_label = val.replace('_', ' ').lower().capitalize()
+        return self.apply_filter(
+            dst='type',
+            src=src,
+            val=type_label
+        )
+
+    def filter_geom(self, src, val):
+        geom = GEOSGeometry(str(val))
+        geom.transform(settings.SRID)
+        return geom
+
+    def filter_attachments(self, src, val):
+        translation_src = self._get_default_translation_src()
+        illustrations = val
+        rv = []
+        for illustration in illustrations:
+            if not illustration.get('traductionFichiers'):
+                continue
+            rv.append(
+                _prepare_attachment_from_apidae_illustration(illustration, translation_src)
+            )
+        return rv