+++ geotrek/trekking/tests/test_parsers.py
@@ -21,7 +21,8 @@ from geotrek.common.tests.mixins import GeotrekParserTestMixin
 from geotrek.trekking.tests.factories import RouteFactory
 from geotrek.trekking.models import POI, Service, Trek, DifficultyLevel, Route
 from geotrek.trekking.parsers import (
-    TrekParser, GeotrekPOIParser, GeotrekServiceParser, GeotrekTrekParser, ApidaeTrekParser, ApidaeTrekThemeParser
+    TrekParser, GeotrekPOIParser, GeotrekServiceParser, GeotrekTrekParser, ApidaeTrekParser, ApidaeTrekThemeParser,
+    ApidaePOIParser, _prepare_attachment_from_apidae_illustration
 )
 
 
@@ -574,8 +575,43 @@ class ServiceGeotrekParserTests(GeotrekParserTestMixin, TestCase):
         self.assertAlmostEqual(service.geom.y, 6192330.15779677, places=5)
 
 
+def make_dummy_apidae_get(parser_class, test_data_dir, data_filename):
+    """Returns a mocked_get. The mocked_get may return:
+    - `data_filename` content on call to APIDAE API,
+    - GPX data file content (the filename is taken from the request's path),
+    - a dummy jpg content if a jpg file is requested.
+
+    parser_class: the class of the parser instance which will call this mocked_get
+    test_data_dir: the path of the directory containing test data relative to the python project root
+    data_filename: the data to return as get response when the parser calls APIDAE API
+    """
+
+    def dummy_get(url, *args, **kwargs):
+        rv = Mock()
+        rv.status_code = 200
+        if url == parser_class.url:
+            filename = os.path.join(test_data_dir, data_filename)
+            with open(filename, 'r') as f:
+                json_payload = f.read()
+            data = json.loads(json_payload)
+            rv.json = lambda: data
+        else:
+            parsed_url = urlparse(url)
+            url_path = parsed_url.path
+            extension = url_path.split('.')[1]
+            if extension == 'jpg':
+                rv.content = copy(testdata.IMG_FILE)
+            elif extension == 'gpx':
+                filename = os.path.join(test_data_dir, url_path.lstrip('/'))
+                with open(filename, 'r') as f:
+                    gpx = f.read()
+                rv.content = bytes(gpx, 'utf-8')
+        return rv
+
+    return dummy_get
+
+
 class TestApidaeTrekParser(ApidaeTrekParser):
-    warn_on_missing_fields = True
     url = 'https://example.net/fake/api/'
     api_key = 'ABCDEF'
     project_id = 1234
@@ -587,30 +623,11 @@ class ApidaeTrekParserTests(TestCase):
 
     @staticmethod
     def make_dummy_get(apidae_data_file):
-
-        def dummy_get(url, *args, **kwargs):
-            rv = Mock()
-            rv.status_code = 200
-            if url == TestApidaeTrekParser.url:
-                filename = apidae_data_file
-                with open(filename, 'r') as f:
-                    json_payload = f.read()
-                data = json.loads(json_payload)
-                rv.json = lambda: data
-            else:
-                parsed_url = urlparse(url)
-                url_path = parsed_url.path
-                extension = url_path.split('.')[1]
-                if extension == 'jpg':
-                    rv.content = copy(testdata.IMG_FILE)
-                elif extension == 'gpx':
-                    filename = os.path.join('geotrek/trekking/tests/data/apidae_trek_parser', url_path.lstrip('/'))
-                    with open(filename, 'r') as f:
-                        gpx = f.read()
-                    rv.content = bytes(gpx, 'utf-8')
-            return rv
-
-        return dummy_get
+        return make_dummy_apidae_get(
+            parser_class=TestApidaeTrekParser,
+            test_data_dir='geotrek/trekking/tests/data/apidae_trek_parser',
+            data_filename=apidae_data_file
+        )
 
     @classmethod
     def setUpTestData(cls):
@@ -619,8 +636,7 @@ class ApidaeTrekParserTests(TestCase):
     @mock.patch('requests.get')
     def test_trek_is_imported(self, mocked_get):
         RouteFactory(route='Boucle')
-
-        mocked_get.side_effect = self.make_dummy_get('geotrek/trekking/tests/data/apidae_trek_parser/a_trek.json')
+        mocked_get.side_effect = self.make_dummy_get('a_trek.json')
 
         call_command('import', 'geotrek.trekking.tests.test_parsers.TestApidaeTrekParser', verbosity=0)
 
@@ -715,16 +731,14 @@ class ApidaeTrekParserTests(TestCase):
         self.assertIn("Cartoguide en vente à l'Office de Tourisme", trek.gear_fr)
 
         # Import an updated trek
-        mocked_get.side_effect = self.make_dummy_get(
-            'geotrek/trekking/tests/data/apidae_trek_parser/a_trek_with_updated_limit_date.json'
-        )
+        mocked_get.side_effect = self.make_dummy_get('a_trek_with_updated_limit_date.json')
         call_command('import', 'geotrek.trekking.tests.test_parsers.TestApidaeTrekParser', verbosity=0)
 
         self.assertEqual(Attachment.objects.count(), 0)
 
     @mock.patch('requests.get')
     def test_trek_geometry_can_be_imported_from_gpx(self, mocked_get):
-        mocked_get.side_effect = self.make_dummy_get('geotrek/trekking/tests/data/apidae_trek_parser/a_trek.json')
+        mocked_get.side_effect = self.make_dummy_get('a_trek.json')
 
         call_command('import', 'geotrek.trekking.tests.test_parsers.TestApidaeTrekParser', verbosity=0)
 
@@ -739,7 +753,7 @@ class ApidaeTrekParserTests(TestCase):
     @mock.patch('requests.get')
     def test_trek_not_imported_when_multiple_plans(self, mocked_get):
         output_stdout = StringIO()
-        mocked_get.side_effect = self.make_dummy_get('geotrek/trekking/tests/data/apidae_trek_parser/trek_multiple_plans_error.json')
+        mocked_get.side_effect = self.make_dummy_get('trek_multiple_plans_error.json')
 
         call_command('import', 'geotrek.trekking.tests.test_parsers.TestApidaeTrekParser', verbosity=2, stdout=output_stdout)
 
@@ -749,8 +763,7 @@ class ApidaeTrekParserTests(TestCase):
     @mock.patch('requests.get')
     def test_trek_not_imported_when_no_gpx_file(self, mocked_get):
         output_stdout = StringIO()
-        mocked_get.side_effect = self.make_dummy_get(
-            'geotrek/trekking/tests/data/apidae_trek_parser/trek_plan_no_gpx_error.json')
+        mocked_get.side_effect = self.make_dummy_get('trek_plan_no_gpx_error.json')
 
         call_command('import', 'geotrek.trekking.tests.test_parsers.TestApidaeTrekParser', verbosity=2,
                      stdout=output_stdout)
@@ -760,8 +773,7 @@ class ApidaeTrekParserTests(TestCase):
 
     @mock.patch('requests.get')
     def test_trek_linked_entities_are_imported(self, mocked_get):
-        mocked_get.side_effect = self.make_dummy_get(
-            'geotrek/trekking/tests/data/apidae_trek_parser/a_trek.json')
+        mocked_get.side_effect = self.make_dummy_get('a_trek.json')
 
         call_command('import', 'geotrek.trekking.tests.test_parsers.TestApidaeTrekParser', verbosity=0)
 
@@ -778,8 +790,7 @@ class ApidaeTrekParserTests(TestCase):
     def test_trek_theme_with_unknown_id_is_not_imported(self, mocked_get):
         assert 12341234 not in ApidaeTrekParser.typologies_sitra_ids_as_themes
 
-        mocked_get.side_effect = self.make_dummy_get(
-            'geotrek/trekking/tests/data/apidae_trek_parser/trek_with_unknown_theme.json')
+        mocked_get.side_effect = self.make_dummy_get('trek_with_unknown_theme.json')
 
         call_command('import', 'geotrek.trekking.tests.test_parsers.TestApidaeTrekParser', verbosity=0)
 
@@ -788,8 +799,7 @@ class ApidaeTrekParserTests(TestCase):
 
     @mock.patch('requests.get')
     def test_links_to_child_treks_are_set(self, mocked_get):
-        mocked_get.side_effect = self.make_dummy_get(
-            'geotrek/trekking/tests/data/apidae_trek_parser/related_treks.json')
+        mocked_get.side_effect = self.make_dummy_get('related_treks.json')
 
         call_command('import', 'geotrek.trekking.tests.test_parsers.TestApidaeTrekParser', verbosity=0)
 
@@ -801,8 +811,7 @@ class ApidaeTrekParserTests(TestCase):
         self.assertIn(parent_trek, child_trek_2.parents.all())
         self.assertEqual(list(parent_trek.children.values_list('eid', flat=True).all()), ['123124', '123125'])
 
-        mocked_get.side_effect = self.make_dummy_get(
-            'geotrek/trekking/tests/data/apidae_trek_parser/related_treks_updated.json')
+        mocked_get.side_effect = self.make_dummy_get('related_treks_updated.json')
 
         call_command('import', 'geotrek.trekking.tests.test_parsers.TestApidaeTrekParser', verbosity=0)
 
@@ -818,8 +827,7 @@ class ApidaeTrekParserTests(TestCase):
 
     @mock.patch('requests.get')
     def test_links_to_child_treks_are_set_with_changed_order_in_data(self, mocked_get):
-        mocked_get.side_effect = self.make_dummy_get(
-            'geotrek/trekking/tests/data/apidae_trek_parser/related_treks_another_order.json')
+        mocked_get.side_effect = self.make_dummy_get('related_treks_another_order.json')
 
         call_command('import', 'geotrek.trekking.tests.test_parsers.TestApidaeTrekParser', verbosity=0)
 
@@ -833,9 +841,7 @@ class ApidaeTrekParserTests(TestCase):
 
     @mock.patch('requests.get')
     def test_trek_illustration_is_not_imported_on_missing_file_metadata(self, mocked_get):
-        mocked_get.side_effect = self.make_dummy_get(
-            'geotrek/trekking/tests/data/apidae_trek_parser/trek_with_not_complete_illustration.json'
-        )
+        mocked_get.side_effect = self.make_dummy_get('trek_with_not_complete_illustration.json')
         call_command('import', 'geotrek.trekking.tests.test_parsers.TestApidaeTrekParser', verbosity=0)
         self.assertEqual(Attachment.objects.count(), 0)
 
@@ -1139,3 +1145,136 @@ class MakeDurationTests(SimpleTestCase):
 
     def test_it_rounds_output_to_two_decimal_places(self):
         self.assertEqual(Decimal(ApidaeTrekParser._make_duration(duration_in_minutes=20)), Decimal('0.33'))
+
+
+class TestApidaePOIParser(ApidaePOIParser):
+    url = 'https://example.net/fake/api/'
+    api_key = 'ABCDEF'
+    project_id = 1234
+    selection_id = 654321
+
+
+@skipIf(settings.TREKKING_TOPOLOGY_ENABLED, 'Test without dynamic segmentation only')
+class ApidaePOIParserTests(TestCase):
+
+    @staticmethod
+    def make_dummy_get(apidae_data_file):
+        return make_dummy_apidae_get(
+            parser_class=TestApidaePOIParser,
+            test_data_dir='geotrek/trekking/tests/data/apidae_poi_parser',
+            data_filename=apidae_data_file
+        )
+
+    @classmethod
+    def setUpTestData(cls):
+        cls.filetype = FileType.objects.create(type="Photographie")
+
+    @mock.patch('requests.get')
+    def test_POI_is_imported(self, mocked_get):
+        mocked_get.side_effect = self.make_dummy_get('a_poi.json')
+
+        call_command('import', 'geotrek.trekking.tests.test_parsers.TestApidaePOIParser', verbosity=0)
+
+        self.assertEqual(POI.objects.count(), 1)
+        poi = POI.objects.all().first()
+        self.assertEqual(poi.name_fr, 'Un point d\'intérêt')
+        self.assertEqual(poi.name_en, 'A point of interest')
+        self.assertEqual(poi.description_fr, 'La description courte en français.')
+        self.assertEqual(poi.description_en, 'The short description in english.')
+
+        self.assertEqual(poi.geom.srid, settings.SRID)
+        self.assertAlmostEqual(poi.geom.coords[0], 729136.5, delta=0.1)
+        self.assertAlmostEqual(poi.geom.coords[1], 6477050.1, delta=0.1)
+
+        self.assertEqual(Attachment.objects.count(), 1)
+        photo = Attachment.objects.first()
+        self.assertEqual(photo.author, 'The author of the picture')
+        self.assertEqual(photo.legend, 'The legend of the picture')
+        self.assertEqual(photo.attachment_file.size, len(testdata.IMG_FILE))
+        self.assertEqual(photo.title, 'The title of the picture')
+
+        self.assertEqual(poi.type.label_en, 'Patrimoine culturel')
+        self.assertEqual(poi.type.label_fr, None)
+
+    @mock.patch('requests.get')
+    def test_trek_illustration_is_not_imported_on_missing_file_metadata(self, mocked_get):
+        mocked_get.side_effect = self.make_dummy_get('poi_with_not_complete_illustration.json')
+        call_command('import', 'geotrek.trekking.tests.test_parsers.TestApidaePOIParser', verbosity=0)
+        self.assertEqual(Attachment.objects.count(), 0)
+
+
+class PrepareAttachmentFromIllustrationTests(TestCase):
+
+    def setUp(self):
+        self.illustration = {
+            'nom': {
+                'libelleEn': 'The title of the picture'
+            },
+            'legende': {
+                'libelleEn': 'The legend of the picture'
+            },
+            'copyright': {
+                'libelleEn': 'The author of the picture'
+            },
+            'traductionFichiers': [
+                {
+                    'url': 'https://example.net/a_picture.jpg'
+                }
+            ]
+        }
+
+    def test_given_full_illustration_it_returns_attachment_info(self):
+        expected_result = (
+            'https://example.net/a_picture.jpg',
+            'The legend of the picture',
+            'The author of the picture',
+            'The title of the picture'
+        )
+        self.assertEqual(
+            _prepare_attachment_from_apidae_illustration(self.illustration, 'libelleEn'),
+            expected_result
+        )
+
+    def test_it_returns_empty_strings_for_missing_info(self):
+        del self.illustration['legende']
+        del self.illustration['copyright']
+        del self.illustration['nom']
+        expected_result = (
+            'https://example.net/a_picture.jpg',
+            '',
+            '',
+            ''
+        )
+        self.assertEqual(
+            _prepare_attachment_from_apidae_illustration(self.illustration, 'libelleEn'),
+            expected_result
+        )
+
+    def test_it_substitutes_name_to_missing_legend(self):
+        del self.illustration['legende']
+        self.illustration['nom'] = {'libelleEn': 'The title of the picture which will also be the legend'}
+        expected_result = (
+            'https://example.net/a_picture.jpg',
+            'The title of the picture which will also be the legend',
+            'The author of the picture',
+            'The title of the picture which will also be the legend'
+        )
+        self.assertEqual(
+            _prepare_attachment_from_apidae_illustration(self.illustration, 'libelleEn'),
+            expected_result
+        )
+
+    def test_it_returns_empty_strings_if_translation_not_found(self):
+        self.illustration['legende'] = {}
+        self.illustration['copyright'] = {}
+        self.illustration['nom'] = {}
+        expected_result = (
+            'https://example.net/a_picture.jpg',
+            '',
+            '',
+            ''
+        )
+        self.assertEqual(
+            _prepare_attachment_from_apidae_illustration(self.illustration, 'libelleEn'),
+            expected_result
+        )